moved config.py inside the package
need to implement passing the main parameters for general config in the parser
